backend_api/requirements.txt
django
djangorestframework
django-cors-headers
django-rest-passwordreset
pymysql
djangorestframework-simplejwt
Pillow
gunicorn
dj-database-url
psycopg2-binary
whitenoise
python-dotenv